get rid of check in CLI
