+++ packages/cli/src/index.tsx
@@ -1,7 +1,6 @@
 import {RenderInternals} from '@remotion/renderer';
 import {benchmarkCommand} from './benchmark';
 import {chalk} from './chalk';
-import {checkNodeVersion} from './check-version';
 import {listCompositionsCommand} from './compositions';
 import {overrideRemotion} from './config/index';
 import {determineFinalImageFormat} from './determine-image-format';
@@ -45,8 +44,6 @@ export const cli = async () => {
 	}
 
 	const remotionRoot = findRemotionRoot();
-	// To check node version and to warn if node version is <12.10.0
-	checkNodeVersion();
 	if (command !== VERSIONS_COMMAND) {
 		await validateVersionsBeforeCommand(remotionRoot);
 	}
